% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mluFVest.R
\name{mluFVest}
\alias{mluFVest}
\title{MLU Fitted Value Estimation}
\usage{
mluFVest(
  model,
  Di,
  Xi,
  Yi,
  Dnewi,
  Xnewi,
  Ynewi,
  Dj = NULL,
  Xj = NULL,
  Yj = NULL,
  Dnewj = NULL,
  Xnewj = NULL,
  Ynewj = NULL,
  f,
  ML = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Torch", "Logit_lasso", "OLS",
    "NLLS_exp", "loglin", "SL", "OLSensemble"),
  shape = c("triangle", "square"),
  polynomial.Lasso = 1,
  polynomial.Ridge = 1,
  polynomial.Logit_lasso = 1,
  polynomial.OLS = 1,
  polynomial.NLLS_exp = 1,
  polynomial.loglin = 1,
  coefs = NULL,
  subsample = NULL
)
}
\arguments{
\item{model}{A model object from mlumodest.}

\item{Di}{Treatment assignment vector for training data (observations i).}

\item{Xi}{Covariate dataframe or matrix for training data (observations i).}

\item{Yi}{Outcome vector for training data (observations i).}

\item{Dnewi}{Treatment assignment vector for new data (observations i).}

\item{Xnewi}{Covariate dataframe or matrix for new data (observations i).}

\item{Ynewi}{Outcome vector for new data (observations i).}

\item{Dj}{Treatment assignment vector for training data (observations j, for "square" mode).}

\item{Xj}{Covariate dataframe or matrix for training data (observations j, for "square" mode).}

\item{Yj}{Outcome vector for training data (observations j, for "square" mode).}

\item{Dnewj}{Treatment assignment vector for new data (observations j, for "square" mode).}

\item{Xnewj}{Covariate dataframe or matrix for new data (observations j, for "square" mode).}

\item{Ynewj}{Outcome vector for new data (observations j, for "square" mode).}

\item{f}{Function of Yi and Yj defining the dependent variable.}

\item{ML}{String vector specifying which machine learners to use.}

\item{shape}{Either "triangle" (within-sample pairs: i < j from same dataset) or
"square" (between-sample pairs: all i,j combinations from two datasets).}

\item{polynomial.Lasso}{Degree of polynomial to be fitted when using Lasso. Default is 1.}

\item{polynomial.Ridge}{Degree of polynomial to be fitted when using Ridge. Default is 1.}

\item{polynomial.Logit_lasso}{Degree of polynomial to be fitted when using Logit_lasso. Default is 1.}

\item{polynomial.OLS}{Degree of polynomial to be fitted when using OLS. Default is 1.}

\item{polynomial.NLLS_exp}{Degree of polynomial to be fitted when using NLLS_exp. Default is 1.}

\item{polynomial.loglin}{Degree of polynomial to be fitted when using loglin. Default is 1.}

\item{coefs}{Optimal coefficients for OLSensemble, computed in mlumodest. Default is NULL.}

\item{subsample}{Either NULL (use all data), a proportion between 0 and 1, or an integer number of observations to randomly subsample before creating pairs. Applies to both i and j observations separately. Default is NULL.}
}
\value{
Fitted values from ML::FVest().
}
\description{
Estimates fitted values for pairwise ML models with treatment assignment.
Supports two modes: "triangle" for within-group pairs and "square" for
between-group pairs. Uses symmetric features: levels (sum) and distances (absolute differences).
}
